UI System(Iteration 1):   -> UI Base Class   -> UI Manager   -> UI Panel

+++ src/engine/camera/AuroraCam.cpp
@@ -40,7 +40,7 @@ namespace tenshi
             f32 _strength = m_ShakeAmount * (1 - ((f32)m_CurrentShakeFrame / (f32)m_ShakeFrames));
 
             m_Offset.x = cosf((f32)GetTime() * 90.0f) * _strength;
-            m_Offset.y = cos((f32)GetTime() * 180.0f) * _strength;
+            m_Offset.y = cosf((f32)GetTime() * 180.0f) * _strength;
 
             ++m_CurrentShakeFrame;
